requirements.txt
flask
torch
torchvision
pillow
flask-cors
